+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Added support for shell autocompletion. ([#65](https://gitlab.eclipse.org/eclipsefdn/security/otterdog/-/issues/65))
 
+### Fixed
+
+- Fixed retrieving repository secrets for temporary private clone repositories.
+
 ## [0.2.0.dev] - unreleased
 
 ### Added